+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * All tables which come from plugin now prefixed as defined by config, #200
 * Now editor styles also adapt to custom styles defined in "Design" tag (e.g. font family, font size, etc), #195
 * Fixed issue when some themes have special styles for `<p>` or `<ul>` tag which caused comment text to be weirdly formatted, #202
+* Fixed issue when user was not moved to the editor on reply/edit/update comment action, #191 
 
 **Other:**